requirements.txt
gitpython
markdown
weasyprint